+++ src/components/mentionRankCard/mentionRankCard.module.css
@@ -1,3 +1,3 @@
-.graphPar {
+.graphBar {
   width: 200px;
 }
