+++ src/pages/Dashboard.js
@@ -1,8 +1,3 @@
-import MentionRankCard from '../components/mentionRankCard/MentionRankCard'
 export default function Dashboard() {
-  return (
-    <>
-      <MentionRankCard />
-    </>
-  )
+  return <></>
 }
